fix: resolve all GCC 11 conversion warnings for Linux shared library build (Task 12.2)
- Fixed size_t to double conversions with static_cast<double>()
- Fixed chrono duration count() to double conversions
- Fixed long double to double conversions in magnitude() calls
- Fixed uint32_t to uint16_t narrowing conversions
- Fixed shadow warnings (time_t renamed to time_val)
- Fixed duplicated-branches warning in solar_system.cpp
- Fixed logical-op warning in web_server.cpp (EAGAIN || EWOULDBLOCK)
- Removed std::vformat/std::make_format_args (not in GCC 11)
- Added missing <filesystem> include in network_mock.cpp
- Added SOLAR_UTILS_API export macro to save_to_file function
- Added solar_jpl to simple_demo link libraries for shared builds
- Added .creation_date = std::nullopt to CelestialBody::Properties

Files modified across:
- lib/solar_core/src/* (multiple files)
- lib/solar_jpl/src/data_validator.cpp
- lib/solar_utils/src/* (multiple files)
- lib/solar_test/src/* (mocks, reporters, formatters, benchmarks)
- apps/* (solar_system, web, realtime, launcher variants)
- examples/* (simple_demo, modern_demo)

Linux shared library build validated:
- SONAME correctly set (libsolar_*.so.4)
- RPATH resolution working ($ORIGIN/../lib)
- Symbol visibility configured
- Installation successful with proper symlinks

+++ apps/solar_system_launcher/error_recovery_enhanced_launcher.cpp
@@ -795,7 +795,7 @@ class ErrorRecoveryLauncherUI {
 
     if (error_stats.total_errors > 0) {
       double recovery_rate = static_cast<double>(error_stats.successful_recoveries) /
-                            (error_stats.successful_recoveries + error_stats.failed_recoveries) * 100.0;
+                            static_cast<double>(error_stats.successful_recoveries + error_stats.failed_recoveries) * 100.0;
       std::cout << "  Recovery Success Rate: " << std::fixed << std::setprecision(1)
                 << recovery_rate << "%\n";
     }
